feat: update winners table
