+++ src/enums/index.ts
@@ -8,7 +8,9 @@ export enum ClientMessageTypesEnum {
   START_GAME = 'start_game',
   TURN = 'turn',
   ATTACK = 'attack',
-  RANDOM_ATTACK = 'randomAttack'
+  RANDOM_ATTACK = 'randomAttack',
+  FINISH = 'finish',
+  UPDATE_WINNERS = 'update_winners',
 }
 
 export enum ErrorMessagesEnum {
