+++ src/helpers/createShotResponse.ts
