+++ src/helpers/createStartGameRes.ts
