+++ src/usersDB/index.ts
@@ -1,11 +1,10 @@
 import { Room } from 'Room';
-import { IUser } from 'types';
-import { User } from 'User/User';
+import { IUser, IWinData } from 'types';
 
 export const users: Array<IUser> = [];
-users.push(new User('admin', 'admin'));
-users.push(new User('newAdmin', 'newAdmin'));
 
 export const rooms: Array<Room> = [];
 
 export const socketsUser = new Map<WebSocket, string>();
+
+export const winnersTable = new Map<string, IWinData>();
