+++ src/utils/index.ts
@@ -1,25 +1,9 @@
 import { ClientMessageTypesEnum } from 'enums';
 import { RequestHandler } from 'RequestHandler/RequestHandler';
-import { IAddUserData, IAttackData, IClientRequest, IRandomAttackData, IShipsData, IUser, IWSRegResponse } from 'types';
+import { IAddUserData, IAttackData, IClientRequest, IRandomAttackData, IShipsData, IWSRegData } from 'types';
 import { users } from 'usersDB';
 export * from './parseString';
 export * from './stringifyObj';
-export * from './createGameRes';
-export * from './createUpdateRoomRes';
-export * from './createShotResponse';
-
-export const createRegResponse = ({ type, data, id, isError, errorText}: Omit<IWSRegResponse, 'data'> & {data: IUser, isError: boolean, errorText: string }): IWSRegResponse => {
-  return {
-    type,
-    id,
-    data: {
-      name: data.name,
-      index: data.id,
-      error: isError,
-      errorText
-    }
-  };
-};
 
 const requestHandler = new RequestHandler(users);
 
@@ -28,7 +12,13 @@ export const checkMessageType = async ({ type, data, id}: IClientRequest, ws: We
   if(type === ClientMessageTypesEnum.REG) {
    const response = await requestHandler.handleRegRequest({ type, data, id}, ws);
    const updatedRoomResponse = await requestHandler.getUpdatedRooms();
-   return [response, updatedRoomResponse];
+   const { name, index, error } = response.data as unknown as IWSRegData;
+   if(error) {
+    return [response, updatedRoomResponse];
+   }
+   
+   const updatedTable = await requestHandler.getUpdatedWinnersTable(index, name);
+   return [response, updatedRoomResponse, updatedTable];
   }
 
   if(type === ClientMessageTypesEnum.CREATE_ROOM) {
@@ -49,7 +39,8 @@ export const checkMessageType = async ({ type, data, id}: IClientRequest, ws: We
   }
 
   if(type === ClientMessageTypesEnum.ATTACK) {
-    await requestHandler.makeAShoot(data as unknown as IAttackData);
+    const response = await requestHandler.makeAShoot(data as unknown as IAttackData);
+    return [response];
   }
 
   if(type === ClientMessageTypesEnum.RANDOM_ATTACK) {
